requirements.txt
python-dotenv
gradio
transformers
torch
faiss-cpu
accelerate
bitsandbytes
huggingface_hub
googlesearch-python
langchain-community
langchain-openai
chromadb